all functionalities finished for map page
Added functionality to check the position of a specific robot, returning either the location name or x-y coordinates, with a confirmation button for checking. Enabled setting the initial position of a robot with the option to select from pre-existing locations. Implemented a dropdown menu for retrieving x-y coordinates by location name, with a confirmation button. Added a button to display all locations along with their x-y coordinates. Introduced a feature to retrieve the current map file with a button for automatic download. Implemented map swapping functionality with file selection and a confirmation button. Removed "View Map File" functionality as map details are displayed at the bottom and on robot selection.