Forgot password functionality & Implement Reset Password
The username field is now back in the login flow after sign-up.
Reset password is now an available option and fully functional.
Forgot password is still a work in progress. Need access to the robonavapp email to enable sending tokens from it.
Backend is fully set up for forgot password functionality.