Added deleted robot, create task and delete task
Added delete robot, create task and delete task. Delete robot ensures there's no instruction on either accept or order status before deleting. Create task creates the outline for the task that later gets instructions queued to it and started. It is created using name and robot id assigned to it. Lastly delete task checks to see if the task is in the complete status if its not then then we need to stop the navigation of the started instruction and delete the other ordered instructions, then delete the task. Made a few assumptions here, we can resort to just waiting for the task to finish.